fix(upload) : adapt click
